Removing an unneccessary mixin
