+++ src/brushes/pencil-eraser-brush.js
@@ -1,10 +1,23 @@
-const MaskedBrushRenderer = require('../masked-brush-renderer');
 const MaskedPath = require('../masked-path');
 
 /**
  * A pencil brush that erases from the canvas
  */
-const PencilEraserBrush = fabric.util.createClass(fabric.PencilBrush, MaskedBrushRenderer, {
+const PencilEraserBrush = fabric.util.createClass(fabric.PencilBrush, {
+  /**
+   * Override the _render method to composite differently than the default pencil brush to allow
+   * eraser previews
+   * @return {void}
+   */
+  _render: function () {
+    var background = this.canvas.wrapperEl.previousElementSibling;
+    var backgroundPattern = this.canvas.contextTop.createPattern(background, 'no-repeat');
+
+    this.canvas.contextTop.strokeStyle = backgroundPattern;
+
+    this.callSuper('_render');
+  },
+
   /**
    * An override of the pencil brush's `createPath` method, to that it uses our
    * custom `MaskedPath`, rather than the default `fabric.Path`. This allows us
